+++ src/ElectroMagn/LaserEnvelope3D.cpp
@@ -75,7 +75,7 @@ LaserEnvelope3D::LaserEnvelope3D( LaserEnvelope *envelope, Patch *patch, Params 
 }
 
 
-void LaserEnvelope3D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
+void LaserEnvelope3D::initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields )
 {
     cField3D *A3D          = static_cast<cField3D *>( A_ );
     cField3D *A03D         = static_cast<cField3D *>( A0_ );
@@ -199,6 +199,66 @@ void LaserEnvelope3D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
         } // end y loop
     } // end x loop
     
+}
+
+void LaserEnvelope3D::injectEnvelopeFromXmin( Patch *patch, Params &params, double time_dual )
+{
+
+    cField3D *A3D                 = static_cast<cField3D *>( A_ );
+    cField3D *A03D                = static_cast<cField3D *>( A0_ );
+
+    double t                      = time_dual;          // x-ct     , t=0
+    double t_previous_timestep    = time_dual-timestep; // x-c(t-dt), t=0
+
+    vector<double> position( 2, 0 );
+    double pos1 = cell_length[1]*( ( double )( patch->getCellStartingGlobalIndex( 1 ) )+( A_->isDual( 1 )?-0.5:0. ) );
+    double pos2 = cell_length[2]*( ( double )( patch->getCellStartingGlobalIndex( 2 ) )+( A_->isDual( 2 )?-0.5:0. ) );
+
+    // oversize
+    int oversize_                 = params.oversize[0];
+
+    bool inject_envelope_from_this_patch = ( patch->isBoundary(0) ) && (  patch->isXmin() );
+    // Impose the envelope value for x=0 at time t and t-dt 
+    if ( inject_envelope_from_this_patch ){
+        position[0] = pos1;
+        for( unsigned int j=0 ; j<A_->dims_[1] ; j++ ) {
+            position[1] = pos2;
+            for( unsigned int k=0 ; k<A_->dims_[2] ; k++ ) {
+                ( *A3D  )( oversize_-1, j, k ) = profile_->complexValueAt( position, t );
+                ( *A03D )( oversize_-1, j, k ) = profile_->complexValueAt( position, t_previous_timestep );
+                position[1] += cell_length[2];
+            } // end j loop
+            position[0] += cell_length[1];
+        } // end k loop
+        
+        if (envelope_solver=="explicit_reduced_dispersion"){
+            // This solver needs another point on the x direction, for one timestep, as initial condition. 
+            // This value will be found by locally solving the envelope wave equation.
+          
+            // Adapting the approach by C. Benedetti described in 
+            // F. Massimo et al., PPCF 2025 https://doi.org/10.1088/1361-6587/addc97,
+            // the second order derivatives in time and longitudinal coordinate are neglected
+            // under the paraxial and slowly varying envelope approximation.
+          
+            // The resulting envelope equation, that will be discretized with first order derivatives, is 
+            // \nabla^2_\perp A + 2*i*(dA/dx+dA/dt) = 0.
+            
+            // from A^n_i and A^(n-1)_i, we find A^n_{i-1}
+            for( unsigned int j=1 ; j<A_->dims_[1]-1 ; j++ ) {
+                for( unsigned int k=1 ; k<A_->dims_[2]-1 ; k++ ) {
+                    // A^n_{i-1} = A^n_i 
+                    ( *A3D  )( oversize_-2, j, k )  = ( *A3D  )( oversize_-1, j, k );
+                    // A^n_{i-1}+= dx/dt*(A^{n+1}_i-A^{n}_i), until here it is like an upwind scheme for the advection equation
+                    ( *A3D  )( oversize_-2, j, k ) += cell_length[0]/timestep* ( ( *A3D  )( oversize_-1, j  , k   )-   ( *A03D)( oversize_-1, j, k ));
+                    // A^n_{i-1}+=dl/(2i) * (\nabla^2_\perp A)|^n_i
+                    ( *A3D  )( oversize_-2, j, k ) += (cell_length[0]/2./i1) * ( ( *A3D  )( oversize_-1, j-1, k   )-2.*( *A3D )( oversize_-1, j, k )+( *A3D )( oversize_-1, j+1, k   ) )*one_ov_dy_sq; // y part
+                    ( *A3D  )( oversize_-2, j, k ) += (cell_length[0]/2./i1) * ( ( *A3D  )( oversize_-1, j  , k-1 )-2.*( *A3D )( oversize_-1, j, k )+( *A3D )( oversize_-1, j  , k+1 ) )*one_ov_dz_sq; // z part
+                } // end k loop 
+            } // end j loop
+            
+        }  // end if envelope_solver=="explicit_reduced_dispersion"
+    }
+      
 }