+++ src/Python/pyprofiles.py
@@ -485,22 +485,32 @@ def LaserPlanar1D( box_side="xmin", a0=1., omega=1.,
     )
 
 def LaserEnvelopePlanar1D( a0=1., omega=1., time_envelope=tconstant(),
-        envelope_solver = "explicit",Envelope_boundary_conditions = [["reflective"]],
+        envelope_solver = "explicit",box_side = "inside",Envelope_boundary_conditions = [["reflective"]],
         polarization_phi = 0.,ellipticity = 0.):
     from numpy import vectorize, sqrt
 
-    def space_time_envelope(x,t):
+    def spatial_envelope1D(x):
         polarization_amplitude_factor = 1/sqrt(1.+ellipticity**2)
-        return (a0*polarization_amplitude_factor) * complex( vectorize(time_envelope)(t) )
+        return (a0*polarization_amplitude_factor)
+    
+    if (box_side=="inside"):
+        def envelope_profile(x,t):
+            return spatial_envelope1D(x)*complex( vectorize(time_envelope)(t) )
+    elif (box_side=="xmin"):
+        def envelope_profile(t):
+            return spatial_envelope1D(0)*complex( vectorize(time_envelope)(t) )
+    else:
+        print("LaserEnvelope error: box_side must be either 'inside' or 'xmin'. ")
 
     # Create Laser Envelope
     LaserEnvelope(
-        omega               = omega,
-        envelope_profile    = space_time_envelope,
-        envelope_solver     = envelope_solver,
+        omega                        = omega,
+        envelope_profile             = envelope_profile,
+        envelope_solver              = envelope_solver,
+        box_side                     = box_side,
         Envelope_boundary_conditions = Envelope_boundary_conditions,
-        polarization_phi    = polarization_phi,
-        ellipticity         = ellipticity
+        polarization_phi             = polarization_phi,
+        ellipticity                  = ellipticity
     )
 
 
@@ -569,13 +579,13 @@ def LaserGaussian2D( box_side="xmin", a0=1., omega=1., focus=None, waist=3., inc
     )
 
 def LaserEnvelopeGaussian2D( a0=1., omega=1., focus=None, waist=3., time_envelope=tconstant(),
-        envelope_solver = "explicit",Envelope_boundary_conditions = [["reflective"]],
+        envelope_solver = "explicit",box_side = "inside",Envelope_boundary_conditions = [["reflective"]],
         polarization_phi = 0.,ellipticity = 0.):
     import cmath
     from numpy import exp, sqrt, arctan, vectorize
     assert len(focus)==2, "LaserEnvelopeGaussian2D: focus must be a list of length 2."
 
-    def gaussian_beam_with_temporal_profile(x,y,t):
+    def gaussian_beam2D(x,y):
         polarization_amplitude_factor = 1/sqrt(1.+ellipticity**2)
         Zr = omega * waist**2/2.
         w  = sqrt(1./(1.+   ( (x-focus[0])/Zr  )**2 ) )
@@ -584,17 +594,26 @@ def LaserEnvelopeGaussian2D( a0=1., omega=1., focus=None, waist=3., time_envelop
         exponential_with_total_phase = exp(1j*(phase-arctan( (x-focus[0])/Zr )))
         invWaist2 = (w/waist)**2
         spatial_amplitude = a0 *polarization_amplitude_factor * sqrt(w) * exp( -invWaist2*(y-focus[1])**2)
-        space_time_envelope = spatial_amplitude * vectorize(time_envelope)(t)
-        return space_time_envelope * exponential_with_total_phase
+        return spatial_amplitude * exponential_with_total_phase
+        
+    if (box_side=="inside"):
+        def envelope_profile(x,y,t):
+            return gaussian_beam2D(x,y)*vectorize(time_envelope)(t)
+    elif (box_side=="xmin"):
+        def envelope_profile(y,t):
+            return gaussian_beam2D(0,y)*vectorize(time_envelope)(t)
+    else:
+        print("LaserEnvelope error: box_side must be either 'inside' or 'xmin'. ")
 
     # Create Laser Envelope
     LaserEnvelope(
-        omega               = omega,
-        envelope_profile    = gaussian_beam_with_temporal_profile,
-        envelope_solver     = envelope_solver,
+        omega                        = omega,
+        envelope_profile             = envelope_profile,
+        envelope_solver              = envelope_solver,
+        box_side      = box_side,
         Envelope_boundary_conditions = Envelope_boundary_conditions,
-        polarization_phi    = polarization_phi,
-        ellipticity         = ellipticity
+        polarization_phi             = polarization_phi,
+        ellipticity                  = ellipticity
     )
 
 def LaserGaussian3D( box_side="xmin", a0=1., omega=1., focus=None, waist=3., incidence_angle=[0.,0.],
@@ -666,13 +685,13 @@ def LaserGaussian3D( box_side="xmin", a0=1., omega=1., focus=None, waist=3., inc
     )
 
 def LaserEnvelopeGaussian3D( a0=1., omega=1., focus=None, waist=3., time_envelope=tconstant(),
-        envelope_solver = "explicit",Envelope_boundary_conditions = [["reflective"]],
+        envelope_solver = "explicit",Envelope_boundary_conditions = [["reflective"]], box_side = "inside",
         polarization_phi = 0.,ellipticity = 0.):
     import cmath
     from numpy import exp, sqrt, arctan, vectorize
     assert len(focus)==3, "LaserEnvelopeGaussian3D: focus must be a list of length 3."
     
-    def gaussian_beam_with_temporal_profile(x,y,z,t):
+    def gaussian_beam3D(x,y,z):
         polarization_amplitude_factor = 1/sqrt(1.+ellipticity**2)
         Zr = omega * waist**2/2.
         w  = sqrt(1./(1.+   ( (x-focus[0])/Zr  )**2 ) )
@@ -681,17 +700,26 @@ def LaserEnvelopeGaussian3D( a0=1., omega=1., focus=None, waist=3., time_envelop
         exponential_with_total_phase = exp(1j*(phase-arctan( (x-focus[0])/Zr )))
         invWaist2 = (w/waist)**2
         spatial_amplitude = a0*polarization_amplitude_factor* w * exp( -invWaist2*(  (y-focus[1])**2 + (z-focus[2])**2 )  )
-        space_time_envelope = spatial_amplitude * vectorize(time_envelope)(t)
-        return space_time_envelope * exponential_with_total_phase
+        return spatial_amplitude * exponential_with_total_phase
+    
+    if (box_side=="inside"):
+        def envelope_profile(x,y,z,t):
+            return gaussian_beam3D(x,y,z)*vectorize(time_envelope)(t)
+    elif (box_side=="xmin"):
+        def envelope_profile(y,z,t):
+            return gaussian_beam3D(0,y,z)*vectorize(time_envelope)(t)
+    else:
+        print("LaserEnvelope error: box_side must be either 'inside' or 'xmin'. ")
 
     # Create Laser Envelope
     LaserEnvelope(
-        omega               = omega,
-        envelope_profile    = gaussian_beam_with_temporal_profile,
-        envelope_solver     = envelope_solver,
+        omega                        = omega,
+        envelope_profile             = envelope_profile,
+        envelope_solver              = envelope_solver,
+        box_side                     = box_side,
         Envelope_boundary_conditions = Envelope_boundary_conditions,
-        polarization_phi    = polarization_phi,
-        ellipticity         = ellipticity
+        polarization_phi             = polarization_phi,
+        ellipticity                  = ellipticity
     )
 
 
@@ -729,7 +757,7 @@ def LaserGaussianAM( box_side="xmin", a0=1., omega=1., focus=None, waist=3.,
 
 
 def LaserEnvelopeGaussianAM( a0=1., omega=1., focus=None, waist=3., time_envelope=tconstant(),
-        envelope_solver = "explicit",Envelope_boundary_conditions = [["reflective"]],
+        envelope_solver = "explicit",box_side = "inside",Envelope_boundary_conditions = [["reflective"]],
         Env_pml_sigma_parameters = [[0.90,2],[10.0,2],[10.0,2]],
         Env_pml_kappa_parameters = [[1.00,1.00,2],[1.00,1.00,2],[1.00,1.00,2]],
         Env_pml_alpha_parameters = [[0.90,0.90,1],[0.75,0.75,1],[0.75,0.75,1]],
@@ -742,7 +770,7 @@ def LaserEnvelopeGaussianAM( a0=1., omega=1., focus=None, waist=3., time_envelop
     elif (len(focus)==2):
         print("WARNING: deprecated focus in LaserEnvelopeGaussianAM should be a list of length 1")
 
-    def gaussian_beam_with_temporal_profile(x,r,t):
+    def gaussian_beamAM(x,r):
         polarization_amplitude_factor = 1/sqrt(1.+ellipticity**2)
         Zr = omega * waist**2/2.
         w  = sqrt(1./(1.+   ( (x-focus[0])/Zr  )**2 ) )
@@ -751,20 +779,29 @@ def LaserEnvelopeGaussianAM( a0=1., omega=1., focus=None, waist=3., time_envelop
         exponential_with_total_phase = exp(1j*(phase-arctan( (x-focus[0])/Zr )))
         invWaist2 = (w/waist)**2
         spatial_amplitude = a0 * polarization_amplitude_factor * w * exp( -invWaist2*(  r**2  ) )
-        space_time_envelope = spatial_amplitude * vectorize(time_envelope)(t)
-        return space_time_envelope * exponential_with_total_phase
-
+        return spatial_amplitude  * exponential_with_total_phase
+        
+    if (box_side=="inside"):
+        def envelope_profile(x,r,t):
+            return gaussian_beamAM(x,r)*vectorize(time_envelope)(t)
+    elif (box_side=="xmin"):
+        def envelope_profile(r,t):
+            return gaussian_beamAM(0,r)*vectorize(time_envelope)(t)
+    else:
+        print("LaserEnvelope error: box_side must be either 'inside' or 'xmin'. ")
+            
     # Create Laser Envelope
     LaserEnvelope(
-        omega               = omega,
-        envelope_profile    = gaussian_beam_with_temporal_profile,
-        envelope_solver     = envelope_solver,
+        omega                        = omega,
+        envelope_profile             = envelope_profile,
+        envelope_solver              = envelope_solver,
+        box_side                     = box_side,
         Envelope_boundary_conditions = Envelope_boundary_conditions,
-        Env_pml_sigma_parameters = Env_pml_sigma_parameters,
-        Env_pml_kappa_parameters = Env_pml_kappa_parameters,
-        Env_pml_alpha_parameters = Env_pml_alpha_parameters,
-        polarization_phi    = polarization_phi,
-        ellipticity         = ellipticity
+        Env_pml_sigma_parameters     = Env_pml_sigma_parameters,
+        Env_pml_kappa_parameters     = Env_pml_kappa_parameters,
+        Env_pml_alpha_parameters     = Env_pml_alpha_parameters,
+        polarization_phi             = polarization_phi,
+        ellipticity                  = ellipticity
     )
 
 # Define the tools for the propagation of a laser profile